fix(examples): relocate harness-endurance benchmark to ai/examples (#13076)
Top-level examples/ is body-only (webpack build-all consumes it); the AI
harness falsifier + its vanilla comparator broke build-all there. Moved to
ai/examples/ (served by the dev-server's process.cwd() static root, so the
e2e still works) + relocated the unit specs to the canonical mirror. Verified:
e2e 4/4, units 16/16.
