+++ ai/examples/harnessEndurance/comparator/markdownBlocks.mjs
