+++ ai/examples/harnessEndurance/neo/index.html
@@ -6,6 +6,6 @@
     <title>Harness Endurance Benchmark — Subject A (Neo)</title>
 </head>
 <body>
-    <script src="../../../src/MicroLoader.mjs" type="module"></script>
+    <script src="../../../../src/MicroLoader.mjs" type="module"></script>
 </body>
 </html>
