+++ ai/examples/harnessEndurance/neo/neo-config.json
@@ -1,0 +1,7 @@
+{
+    "appPath"    : "ai/examples/harnessEndurance/neo/app.mjs",
+    "basePath"   : "../../../../",
+    "environment": "development",
+    "mainPath"   : "./Main.mjs",
+    "useAiClient": true
+}
